@@ -26,7 +26,7 @@ class AnimateCamera extends StatefulWidget {
 }
 
 class AnimateCameraState extends State<AnimateCamera> {
-  GoogleMapController mapController;
+  GoogleMapController? mapController;
 
   void _onMapCreated(GoogleMapController controller) {
     mapController = controller;
@@ -56,7 +56,7 @@ class AnimateCameraState extends State<AnimateCamera> {
               children: <Widget>[
                 TextButton(
                   onPressed: () {
-                    mapController.animateCamera(
+                    mapController?.animateCamera(
                       CameraUpdate.newCameraPosition(
                         const CameraPosition(
                           bearing: 270.0,
@@ -71,7 +71,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.animateCamera(
+                    mapController?.animateCamera(
                       CameraUpdate.newLatLng(
                         const LatLng(56.1725505, 10.1850512),
                       ),
@@ -81,7 +81,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.animateCamera(
+                    mapController?.animateCamera(
                       CameraUpdate.newLatLngBounds(
                         LatLngBounds(
                           southwest: const LatLng(-38.483935, 113.248673),
@@ -95,7 +95,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.animateCamera(
+                    mapController?.animateCamera(
                       CameraUpdate.newLatLngZoom(
                         const LatLng(37.4231613, -122.087159),
                         11.0,
@@ -106,7 +106,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.animateCamera(
+                    mapController?.animateCamera(
                       CameraUpdate.scrollBy(150.0, -225.0),
                     );
                   },
@@ -118,7 +118,7 @@ class AnimateCameraState extends State<AnimateCamera> {
               children: <Widget>[
                 TextButton(
                   onPressed: () {
-                    mapController.animateCamera(
+                    mapController?.animateCamera(
                       CameraUpdate.zoomBy(
                         -0.5,
                         const Offset(30.0, 20.0),
@@ -129,7 +129,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.animateCamera(
+                    mapController?.animateCamera(
                       CameraUpdate.zoomBy(-0.5),
                     );
                   },
@@ -137,7 +137,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.animateCamera(
+                    mapController?.animateCamera(
                       CameraUpdate.zoomIn(),
                     );
                   },
@@ -145,7 +145,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.animateCamera(
+                    mapController?.animateCamera(
                       CameraUpdate.zoomOut(),
                     );
                   },
@@ -153,7 +153,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.animateCamera(
+                    mapController?.animateCamera(
                       CameraUpdate.zoomTo(16.0),
                     );
                   },
