@@ -56,7 +56,7 @@ class MapUiBodyState extends State<MapUiBody> {
   bool _myLocationEnabled = true;
   bool _myTrafficEnabled = false;
   bool _myLocationButtonEnabled = true;
-  GoogleMapController _controller;
+  late GoogleMapController _controller;
   bool _nightMode = false;
 
   @override
@@ -249,10 +249,9 @@ class MapUiBodyState extends State<MapUiBody> {
     });
   }
 
+  // Should only be called if _isMapCreated is true.
   Widget _nightModeToggler() {
-    if (!_isMapCreated) {
-      return null;
-    }
+    assert(_isMapCreated);
     return TextButton(
       child: Text('${_nightMode ? 'disable' : 'enable'} night mode'),
       onPressed: () {
