@@ -29,8 +29,8 @@ class MarkerIconsBody extends StatefulWidget {
 const LatLng _kMapCenter = LatLng(52.4478, -3.5402);
 
 class MarkerIconsBodyState extends State<MarkerIconsBody> {
-  GoogleMapController controller;
-  BitmapDescriptor _markerIcon;
+  GoogleMapController? controller;
+  BitmapDescriptor? _markerIcon;
 
   @override
   Widget build(BuildContext context) {
@@ -48,7 +48,7 @@ class MarkerIconsBodyState extends State<MarkerIconsBody> {
                 target: _kMapCenter,
                 zoom: 7.0,
               ),
-              markers: _createMarker(),
+              markers: <Marker>{_createMarker()},
               onMapCreated: _onMapCreated,
             ),
           ),
@@ -57,17 +57,19 @@ class MarkerIconsBodyState extends State<MarkerIconsBody> {
     );
   }
 
-  Set<Marker> _createMarker() {
-    // TODO(iskakaushik): Remove this when collection literals makes it to stable.
-    // https://github.com/flutter/flutter/issues/28312
-    // ignore: prefer_collection_literals
-    return <Marker>[
-      Marker(
+  Marker _createMarker() {
+    if (_markerIcon != null) {
+      return Marker(
         markerId: MarkerId("marker_1"),
         position: _kMapCenter,
-        icon: _markerIcon,
-      ),
-    ].toSet();
+        icon: _markerIcon!,
+      );
+    } else {
+      return Marker(
+        markerId: MarkerId("marker_1"),
+        position: _kMapCenter,
+      );
+    }
   }
 
   Future<void> _createMarkerImageFromAsset(BuildContext context) async {
