@@ -25,7 +25,7 @@ class MoveCamera extends StatefulWidget {
 }
 
 class MoveCameraState extends State<MoveCamera> {
-  GoogleMapController mapController;
+  GoogleMapController? mapController;
 
   void _onMapCreated(GoogleMapController controller) {
     mapController = controller;
@@ -55,7 +55,7 @@ class MoveCameraState extends State<MoveCamera> {
               children: <Widget>[
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController?.moveCamera(
                       CameraUpdate.newCameraPosition(
                         const CameraPosition(
                           bearing: 270.0,
@@ -70,7 +70,7 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController?.moveCamera(
                       CameraUpdate.newLatLng(
                         const LatLng(56.1725505, 10.1850512),
                       ),
@@ -80,7 +80,7 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController?.moveCamera(
                       CameraUpdate.newLatLngBounds(
                         LatLngBounds(
                           southwest: const LatLng(-38.483935, 113.248673),
@@ -94,7 +94,7 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController?.moveCamera(
                       CameraUpdate.newLatLngZoom(
                         const LatLng(37.4231613, -122.087159),
                         11.0,
@@ -105,7 +105,7 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController?.moveCamera(
                       CameraUpdate.scrollBy(150.0, -225.0),
                     );
                   },
@@ -117,7 +117,7 @@ class MoveCameraState extends State<MoveCamera> {
               children: <Widget>[
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController?.moveCamera(
                       CameraUpdate.zoomBy(
                         -0.5,
                         const Offset(30.0, 20.0),
@@ -128,7 +128,7 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController?.moveCamera(
                       CameraUpdate.zoomBy(-0.5),
                     );
                   },
@@ -136,7 +136,7 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController?.moveCamera(
                       CameraUpdate.zoomIn(),
                     );
                   },
@@ -144,7 +144,7 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController?.moveCamera(
                       CameraUpdate.zoomOut(),
                     );
                   },
@@ -152,7 +152,7 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController?.moveCamera(
                       CameraUpdate.zoomTo(16.0),
                     );
                   },
