@@ -28,10 +28,10 @@ class PlaceCircleBody extends StatefulWidget {
 class PlaceCircleBodyState extends State<PlaceCircleBody> {
   PlaceCircleBodyState();
 
-  GoogleMapController controller;
+  GoogleMapController? controller;
   Map<CircleId, Circle> circles = <CircleId, Circle>{};
   int _circleIdCounter = 1;
-  CircleId selectedCircle;
+  CircleId? selectedCircle;
 
   // Values when toggling circle color
   int fillColorsIndex = 0;
@@ -62,12 +62,14 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
     });
   }
 
-  void _remove() {
+  void _remove(CircleId circleId) {
     setState(() {
-      if (circles.containsKey(selectedCircle)) {
-        circles.remove(selectedCircle);
+      if (circles.containsKey(circleId)) {
+        circles.remove(circleId);
       }
-      selectedCircle = null;
+      if (circleId == selectedCircle) {
+        selectedCircle = null;
+      }
     });
   }
 
@@ -100,37 +102,37 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
     });
   }
 
-  void _toggleVisible() {
-    final Circle circle = circles[selectedCircle];
+  void _toggleVisible(CircleId circleId) {
+    final Circle circle = circles[circleId]!;
     setState(() {
-      circles[selectedCircle] = circle.copyWith(
+      circles[circleId] = circle.copyWith(
         visibleParam: !circle.visible,
       );
     });
   }
 
-  void _changeFillColor() {
-    final Circle circle = circles[selectedCircle];
+  void _changeFillColor(CircleId circleId) {
+    final Circle circle = circles[circleId]!;
     setState(() {
-      circles[selectedCircle] = circle.copyWith(
+      circles[circleId] = circle.copyWith(
         fillColorParam: colors[++fillColorsIndex % colors.length],
       );
     });
   }
 
-  void _changeStrokeColor() {
-    final Circle circle = circles[selectedCircle];
+  void _changeStrokeColor(CircleId circleId) {
+    final Circle circle = circles[circleId]!;
     setState(() {
-      circles[selectedCircle] = circle.copyWith(
+      circles[circleId] = circle.copyWith(
         strokeColorParam: colors[++strokeColorsIndex % colors.length],
       );
     });
   }
 
-  void _changeStrokeWidth() {
-    final Circle circle = circles[selectedCircle];
+  void _changeStrokeWidth(CircleId circleId) {
+    final Circle circle = circles[circleId]!;
     setState(() {
-      circles[selectedCircle] = circle.copyWith(
+      circles[circleId] = circle.copyWith(
         strokeWidthParam: widths[++widthsIndex % widths.length],
       );
     });
@@ -138,6 +140,7 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
 
   @override
   Widget build(BuildContext context) {
+    final CircleId? selectedId = selectedCircle;
     return Column(
       mainAxisAlignment: MainAxisAlignment.spaceEvenly,
       crossAxisAlignment: CrossAxisAlignment.stretch,
@@ -171,12 +174,15 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
                         ),
                         TextButton(
                           child: const Text('remove'),
-                          onPressed: (selectedCircle == null) ? null : _remove,
+                          onPressed: (selectedId == null)
+                              ? null
+                              : () => _remove(selectedId),
                         ),
                         TextButton(
                           child: const Text('toggle visible'),
-                          onPressed:
-                              (selectedCircle == null) ? null : _toggleVisible,
+                          onPressed: (selectedId == null)
+                              ? null
+                              : () => _toggleVisible(selectedId),
                         ),
                       ],
                     ),
@@ -184,21 +190,21 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
                       children: <Widget>[
                         TextButton(
                           child: const Text('change stroke width'),
-                          onPressed: (selectedCircle == null)
+                          onPressed: (selectedId == null)
                               ? null
-                              : _changeStrokeWidth,
+                              : () => _changeStrokeWidth(selectedId),
                         ),
                         TextButton(
                           child: const Text('change stroke color'),
-                          onPressed: (selectedCircle == null)
+                          onPressed: (selectedId == null)
                               ? null
-                              : _changeStrokeColor,
+                              : () => _changeStrokeColor(selectedId),
                         ),
                         TextButton(
                           child: const Text('change fill color'),
-                          onPressed: (selectedCircle == null)
+                          onPressed: (selectedId == null)
                               ? null
-                              : _changeFillColor,
+                              : () => _changeFillColor(selectedId),
                         ),
                       ],
                     )
