@@ -9,20 +9,6 @@ import 'package:google_maps_flutter/google_maps_flutter.dart';
 
 import 'fake_maps_controllers.dart';
 
-Set<Polygon> _toSet({Polygon p1, Polygon p2, Polygon p3}) {
-  final Set<Polygon> res = Set<Polygon>.identity();
-  if (p1 != null) {
-    res.add(p1);
-  }
-  if (p2 != null) {
-    res.add(p2);
-  }
-  if (p3 != null) {
-    res.add(p3);
-  }
-  return res;
-}
-
 Widget _mapWithPolygons(Set<Polygon> polygons) {
   return Directionality(
     textDirection: TextDirection.ltr,
@@ -34,7 +20,7 @@ Widget _mapWithPolygons(Set<Polygon> polygons) {
 }
 
 List<LatLng> _rectPoints({
-  @required double size,
+  required double size,
   LatLng center = const LatLng(0, 0),
 }) {
   final halfSize = size / 2;
@@ -73,10 +59,10 @@ void main() {
 
   testWidgets('Initializing a polygon', (WidgetTester tester) async {
     final Polygon p1 = Polygon(polygonId: PolygonId("polygon_1"));
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1)));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p1}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.polygonsToAdd.length, 1);
 
     final Polygon initializedPolygon = platformGoogleMap.polygonsToAdd.first;
@@ -89,11 +75,11 @@ void main() {
     final Polygon p1 = Polygon(polygonId: PolygonId("polygon_1"));
     final Polygon p2 = Polygon(polygonId: PolygonId("polygon_2"));
 
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1)));
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1, p2: p2)));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p1}));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p1, p2}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.polygonsToAdd.length, 1);
 
     final Polygon addedPolygon = platformGoogleMap.polygonsToAdd.first;
@@ -107,11 +93,11 @@ void main() {
   testWidgets("Removing a polygon", (WidgetTester tester) async {
     final Polygon p1 = Polygon(polygonId: PolygonId("polygon_1"));
 
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1)));
-    await tester.pumpWidget(_mapWithPolygons(null));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p1}));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.polygonIdsToRemove.length, 1);
     expect(platformGoogleMap.polygonIdsToRemove.first, equals(p1.polygonId));
 
@@ -124,11 +110,11 @@ void main() {
     final Polygon p2 =
         Polygon(polygonId: PolygonId("polygon_1"), geodesic: true);
 
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1)));
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p2)));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p1}));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p2}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.polygonsToChange.length, 1);
     expect(platformGoogleMap.polygonsToChange.first, equals(p2));
 
@@ -141,13 +127,13 @@ void main() {
       polygonId: PolygonId("polygon_1"),
       points: <LatLng>[const LatLng(0.0, 0.0)],
     );
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1)));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p1}));
 
     p1.points.add(const LatLng(1.0, 1.0));
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1)));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p1}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.polygonsToChange.length, 1);
     expect(platformGoogleMap.polygonsToChange.first, equals(p1));
 
@@ -158,16 +144,16 @@ void main() {
   testWidgets("Multi Update", (WidgetTester tester) async {
     Polygon p1 = Polygon(polygonId: PolygonId("polygon_1"));
     Polygon p2 = Polygon(polygonId: PolygonId("polygon_2"));
-    final Set<Polygon> prev = _toSet(p1: p1, p2: p2);
+    final Set<Polygon> prev = <Polygon>{p1, p2};
     p1 = Polygon(polygonId: PolygonId("polygon_1"), visible: false);
     p2 = Polygon(polygonId: PolygonId("polygon_2"), geodesic: true);
-    final Set<Polygon> cur = _toSet(p1: p1, p2: p2);
+    final Set<Polygon> cur = <Polygon>{p1, p2};
 
     await tester.pumpWidget(_mapWithPolygons(prev));
     await tester.pumpWidget(_mapWithPolygons(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformGoogleMap.polygonsToChange, cur);
     expect(platformGoogleMap.polygonIdsToRemove.isEmpty, true);
@@ -177,18 +163,18 @@ void main() {
   testWidgets("Multi Update", (WidgetTester tester) async {
     Polygon p2 = Polygon(polygonId: PolygonId("polygon_2"));
     final Polygon p3 = Polygon(polygonId: PolygonId("polygon_3"));
-    final Set<Polygon> prev = _toSet(p2: p2, p3: p3);
+    final Set<Polygon> prev = <Polygon>{p2, p3};
 
     // p1 is added, p2 is updated, p3 is removed.
     final Polygon p1 = Polygon(polygonId: PolygonId("polygon_1"));
     p2 = Polygon(polygonId: PolygonId("polygon_2"), geodesic: true);
-    final Set<Polygon> cur = _toSet(p1: p1, p2: p2);
+    final Set<Polygon> cur = <Polygon>{p1, p2};
 
     await tester.pumpWidget(_mapWithPolygons(prev));
     await tester.pumpWidget(_mapWithPolygons(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformGoogleMap.polygonsToChange.length, 1);
     expect(platformGoogleMap.polygonsToAdd.length, 1);
@@ -203,32 +189,32 @@ void main() {
     final Polygon p1 = Polygon(polygonId: PolygonId("polygon_1"));
     final Polygon p2 = Polygon(polygonId: PolygonId("polygon_2"));
     Polygon p3 = Polygon(polygonId: PolygonId("polygon_3"));
-    final Set<Polygon> prev = _toSet(p1: p1, p2: p2, p3: p3);
+    final Set<Polygon> prev = <Polygon>{p1, p2, p3};
     p3 = Polygon(polygonId: PolygonId("polygon_3"), geodesic: true);
-    final Set<Polygon> cur = _toSet(p1: p1, p2: p2, p3: p3);
+    final Set<Polygon> cur = <Polygon>{p1, p2, p3};
 
     await tester.pumpWidget(_mapWithPolygons(prev));
     await tester.pumpWidget(_mapWithPolygons(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
-    expect(platformGoogleMap.polygonsToChange, _toSet(p3: p3));
+    expect(platformGoogleMap.polygonsToChange, <Polygon>{p3});
     expect(platformGoogleMap.polygonIdsToRemove.isEmpty, true);
     expect(platformGoogleMap.polygonsToAdd.isEmpty, true);
   });
 
   testWidgets("Update non platform related attr", (WidgetTester tester) async {
     Polygon p1 = Polygon(polygonId: PolygonId("polygon_1"));
-    final Set<Polygon> prev = _toSet(p1: p1);
+    final Set<Polygon> prev = <Polygon>{p1};
     p1 = Polygon(polygonId: PolygonId("polygon_1"), onTap: () => print(2 + 2));
-    final Set<Polygon> cur = _toSet(p1: p1);
+    final Set<Polygon> cur = <Polygon>{p1};
 
     await tester.pumpWidget(_mapWithPolygons(prev));
     await tester.pumpWidget(_mapWithPolygons(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformGoogleMap.polygonsToChange.isEmpty, true);
     expect(platformGoogleMap.polygonIdsToRemove.isEmpty, true);
@@ -238,10 +224,10 @@ void main() {
   testWidgets('Initializing a polygon with points and hole',
       (WidgetTester tester) async {
     final Polygon p1 = _polygonWithPointsAndHole(PolygonId("polygon_1"));
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1)));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p1}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.polygonsToAdd.length, 1);
 
     final Polygon initializedPolygon = platformGoogleMap.polygonsToAdd.first;
@@ -255,11 +241,11 @@ void main() {
     final Polygon p1 = Polygon(polygonId: PolygonId("polygon_1"));
     final Polygon p2 = _polygonWithPointsAndHole(PolygonId("polygon_2"));
 
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1)));
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1, p2: p2)));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p1}));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p1, p2}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.polygonsToAdd.length, 1);
 
     final Polygon addedPolygon = platformGoogleMap.polygonsToAdd.first;
@@ -274,11 +260,11 @@ void main() {
       (WidgetTester tester) async {
     final Polygon p1 = _polygonWithPointsAndHole(PolygonId("polygon_1"));
 
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1)));
-    await tester.pumpWidget(_mapWithPolygons(null));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p1}));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.polygonIdsToRemove.length, 1);
     expect(platformGoogleMap.polygonIdsToRemove.first, equals(p1.polygonId));
 
@@ -291,11 +277,11 @@ void main() {
     final Polygon p1 = Polygon(polygonId: PolygonId("polygon_1"));
     final Polygon p2 = _polygonWithPointsAndHole(PolygonId("polygon_1"));
 
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1)));
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p2)));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p1}));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p2}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.polygonsToChange.length, 1);
     expect(platformGoogleMap.polygonsToChange.first, equals(p2));
 
@@ -310,7 +296,7 @@ void main() {
       points: _rectPoints(size: 1),
       holes: [_rectPoints(size: 0.5)],
     );
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1)));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p1}));
 
     p1.points
       ..clear()
@@ -318,10 +304,10 @@ void main() {
     p1.holes
       ..clear()
       ..addAll([_rectPoints(size: 1)]);
-    await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1)));
+    await tester.pumpWidget(_mapWithPolygons(<Polygon>{p1}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.polygonsToChange.length, 1);
     expect(platformGoogleMap.polygonsToChange.first, equals(p1));
 
@@ -337,19 +323,19 @@ void main() {
       points: _rectPoints(size: 2),
       holes: [_rectPoints(size: 1)],
     );
-    final Set<Polygon> prev = _toSet(p1: p1, p2: p2);
+    final Set<Polygon> prev = <Polygon>{p1, p2};
     p1 = Polygon(polygonId: PolygonId("polygon_1"), visible: false);
     p2 = p2.copyWith(
       pointsParam: _rectPoints(size: 5),
       holesParam: [_rectPoints(size: 2)],
     );
-    final Set<Polygon> cur = _toSet(p1: p1, p2: p2);
+    final Set<Polygon> cur = <Polygon>{p1, p2};
 
     await tester.pumpWidget(_mapWithPolygons(prev));
     await tester.pumpWidget(_mapWithPolygons(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformGoogleMap.polygonsToChange, cur);
     expect(platformGoogleMap.polygonIdsToRemove.isEmpty, true);
@@ -364,7 +350,7 @@ void main() {
       holes: [_rectPoints(size: 1)],
     );
     final Polygon p3 = Polygon(polygonId: PolygonId("polygon_3"));
-    final Set<Polygon> prev = _toSet(p2: p2, p3: p3);
+    final Set<Polygon> prev = <Polygon>{p2, p3};
 
     // p1 is added, p2 is updated, p3 is removed.
     final Polygon p1 = _polygonWithPointsAndHole(PolygonId("polygon_1"));
@@ -372,13 +358,13 @@ void main() {
       pointsParam: _rectPoints(size: 5),
       holesParam: [_rectPoints(size: 3)],
     );
-    final Set<Polygon> cur = _toSet(p1: p1, p2: p2);
+    final Set<Polygon> cur = <Polygon>{p1, p2};
 
     await tester.pumpWidget(_mapWithPolygons(prev));
     await tester.pumpWidget(_mapWithPolygons(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformGoogleMap.polygonsToChange.length, 1);
     expect(platformGoogleMap.polygonsToAdd.length, 1);
@@ -398,20 +384,20 @@ void main() {
       points: _rectPoints(size: 2),
       holes: [_rectPoints(size: 1)],
     );
-    final Set<Polygon> prev = _toSet(p1: p1, p2: p2, p3: p3);
+    final Set<Polygon> prev = <Polygon>{p1, p2, p3};
     p3 = p3.copyWith(
       pointsParam: _rectPoints(size: 5),
       holesParam: [_rectPoints(size: 3)],
     );
-    final Set<Polygon> cur = _toSet(p1: p1, p2: p2, p3: p3);
+    final Set<Polygon> cur = <Polygon>{p1, p2, p3};
 
     await tester.pumpWidget(_mapWithPolygons(prev));
     await tester.pumpWidget(_mapWithPolygons(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
-    expect(platformGoogleMap.polygonsToChange, _toSet(p3: p3));
+    expect(platformGoogleMap.polygonsToChange, <Polygon>{p3});
     expect(platformGoogleMap.polygonIdsToRemove.isEmpty, true);
     expect(platformGoogleMap.polygonsToAdd.isEmpty, true);
   });
