@@ -1,6 +1,7 @@
 // Copyright 2019, the Chromium project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+// @dart=2.9
 
 import 'dart:async';
 import 'dart:io';
@@ -954,10 +955,8 @@ void main() {
     final BitmapDescriptor scaled = await BitmapDescriptor.fromAssetImage(
         imageConfiguration, 'red_square.png',
         mipmaps: false);
-    // ignore: invalid_use_of_visible_for_testing_member
-    expect(mip.toJson()[2], 1);
-    // ignore: invalid_use_of_visible_for_testing_member
-    expect(scaled.toJson()[2], 2);
+    expect((mip.toJson() as List<dynamic>)[2], 1);
+    expect((scaled.toJson() as List<dynamic>)[2], 2);
   });
 
   testWidgets('testTakeSnapshot', (WidgetTester tester) async {
