@@ -9,20 +9,6 @@ import 'package:google_maps_flutter/google_maps_flutter.dart';
 
 import 'fake_maps_controllers.dart';
 
-Set<Circle> _toSet({Circle c1, Circle c2, Circle c3}) {
-  final Set<Circle> res = Set<Circle>.identity();
-  if (c1 != null) {
-    res.add(c1);
-  }
-  if (c2 != null) {
-    res.add(c2);
-  }
-  if (c3 != null) {
-    res.add(c3);
-  }
-  return res;
-}
-
 Widget _mapWithCircles(Set<Circle> circles) {
   return Directionality(
     textDirection: TextDirection.ltr,
@@ -50,10 +36,10 @@ void main() {
 
   testWidgets('Initializing a circle', (WidgetTester tester) async {
     final Circle c1 = Circle(circleId: CircleId("circle_1"));
-    await tester.pumpWidget(_mapWithCircles(_toSet(c1: c1)));
+    await tester.pumpWidget(_mapWithCircles(<Circle>{c1}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.circlesToAdd.length, 1);
 
     final Circle initializedCircle = platformGoogleMap.circlesToAdd.first;
@@ -66,11 +52,11 @@ void main() {
     final Circle c1 = Circle(circleId: CircleId("circle_1"));
     final Circle c2 = Circle(circleId: CircleId("circle_2"));
 
-    await tester.pumpWidget(_mapWithCircles(_toSet(c1: c1)));
-    await tester.pumpWidget(_mapWithCircles(_toSet(c1: c1, c2: c2)));
+    await tester.pumpWidget(_mapWithCircles(<Circle>{c1}));
+    await tester.pumpWidget(_mapWithCircles(<Circle>{c1, c2}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.circlesToAdd.length, 1);
 
     final Circle addedCircle = platformGoogleMap.circlesToAdd.first;
@@ -84,11 +70,11 @@ void main() {
   testWidgets("Removing a circle", (WidgetTester tester) async {
     final Circle c1 = Circle(circleId: CircleId("circle_1"));
 
-    await tester.pumpWidget(_mapWithCircles(_toSet(c1: c1)));
-    await tester.pumpWidget(_mapWithCircles(null));
+    await tester.pumpWidget(_mapWithCircles(<Circle>{c1}));
+    await tester.pumpWidget(_mapWithCircles(<Circle>{}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.circleIdsToRemove.length, 1);
     expect(platformGoogleMap.circleIdsToRemove.first, equals(c1.circleId));
 
@@ -100,11 +86,11 @@ void main() {
     final Circle c1 = Circle(circleId: CircleId("circle_1"));
     final Circle c2 = Circle(circleId: CircleId("circle_1"), radius: 10);
 
-    await tester.pumpWidget(_mapWithCircles(_toSet(c1: c1)));
-    await tester.pumpWidget(_mapWithCircles(_toSet(c1: c2)));
+    await tester.pumpWidget(_mapWithCircles(<Circle>{c1}));
+    await tester.pumpWidget(_mapWithCircles(<Circle>{c2}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.circlesToChange.length, 1);
     expect(platformGoogleMap.circlesToChange.first, equals(c2));
 
@@ -116,11 +102,11 @@ void main() {
     final Circle c1 = Circle(circleId: CircleId("circle_1"));
     final Circle c2 = Circle(circleId: CircleId("circle_1"), radius: 10);
 
-    await tester.pumpWidget(_mapWithCircles(_toSet(c1: c1)));
-    await tester.pumpWidget(_mapWithCircles(_toSet(c1: c2)));
+    await tester.pumpWidget(_mapWithCircles(<Circle>{c1}));
+    await tester.pumpWidget(_mapWithCircles(<Circle>{c2}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.circlesToChange.length, 1);
 
     final Circle update = platformGoogleMap.circlesToChange.first;
@@ -131,16 +117,16 @@ void main() {
   testWidgets("Multi Update", (WidgetTester tester) async {
     Circle c1 = Circle(circleId: CircleId("circle_1"));
     Circle c2 = Circle(circleId: CircleId("circle_2"));
-    final Set<Circle> prev = _toSet(c1: c1, c2: c2);
+    final Set<Circle> prev = <Circle>{c1, c2};
     c1 = Circle(circleId: CircleId("circle_1"), visible: false);
     c2 = Circle(circleId: CircleId("circle_2"), radius: 10);
-    final Set<Circle> cur = _toSet(c1: c1, c2: c2);
+    final Set<Circle> cur = <Circle>{c1, c2};
 
     await tester.pumpWidget(_mapWithCircles(prev));
     await tester.pumpWidget(_mapWithCircles(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformGoogleMap.circlesToChange, cur);
     expect(platformGoogleMap.circleIdsToRemove.isEmpty, true);
@@ -150,18 +136,18 @@ void main() {
   testWidgets("Multi Update", (WidgetTester tester) async {
     Circle c2 = Circle(circleId: CircleId("circle_2"));
     final Circle c3 = Circle(circleId: CircleId("circle_3"));
-    final Set<Circle> prev = _toSet(c2: c2, c3: c3);
+    final Set<Circle> prev = <Circle>{c2, c3};
 
     // c1 is added, c2 is updated, c3 is removed.
     final Circle c1 = Circle(circleId: CircleId("circle_1"));
     c2 = Circle(circleId: CircleId("circle_2"), radius: 10);
-    final Set<Circle> cur = _toSet(c1: c1, c2: c2);
+    final Set<Circle> cur = <Circle>{c1, c2};
 
     await tester.pumpWidget(_mapWithCircles(prev));
     await tester.pumpWidget(_mapWithCircles(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformGoogleMap.circlesToChange.length, 1);
     expect(platformGoogleMap.circlesToAdd.length, 1);
@@ -176,32 +162,32 @@ void main() {
     final Circle c1 = Circle(circleId: CircleId("circle_1"));
     final Circle c2 = Circle(circleId: CircleId("circle_2"));
     Circle c3 = Circle(circleId: CircleId("circle_3"));
-    final Set<Circle> prev = _toSet(c1: c1, c2: c2, c3: c3);
+    final Set<Circle> prev = <Circle>{c1, c2, c3};
     c3 = Circle(circleId: CircleId("circle_3"), radius: 10);
-    final Set<Circle> cur = _toSet(c1: c1, c2: c2, c3: c3);
+    final Set<Circle> cur = <Circle>{c1, c2, c3};
 
     await tester.pumpWidget(_mapWithCircles(prev));
     await tester.pumpWidget(_mapWithCircles(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
-    expect(platformGoogleMap.circlesToChange, _toSet(c3: c3));
+    expect(platformGoogleMap.circlesToChange, <Circle>{c3});
     expect(platformGoogleMap.circleIdsToRemove.isEmpty, true);
     expect(platformGoogleMap.circlesToAdd.isEmpty, true);
   });
 
   testWidgets("Update non platform related attr", (WidgetTester tester) async {
     Circle c1 = Circle(circleId: CircleId("circle_1"));
-    final Set<Circle> prev = _toSet(c1: c1);
+    final Set<Circle> prev = <Circle>{c1};
     c1 = Circle(circleId: CircleId("circle_1"), onTap: () => print("hello"));
-    final Set<Circle> cur = _toSet(c1: c1);
+    final Set<Circle> cur = <Circle>{c1};
 
     await tester.pumpWidget(_mapWithCircles(prev));
     await tester.pumpWidget(_mapWithCircles(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformGoogleMap.circlesToChange.isEmpty, true);
     expect(platformGoogleMap.circleIdsToRemove.isEmpty, true);
