@@ -2,26 +2,26 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'dart:async';
+import 'dart:typed_data';
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter/gestures.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:google_maps_flutter_platform_interface/google_maps_flutter_platform_interface.dart';
-import 'package:plugin_platform_interface/plugin_platform_interface.dart';
-import 'package:mockito/mockito.dart';
-
-class MockGoogleMapsFlutterPlatform extends Mock
-    with MockPlatformInterfaceMixin
-    implements GoogleMapsFlutterPlatform {}
+import 'package:stream_transform/stream_transform.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
-  final platform = MockGoogleMapsFlutterPlatform();
+  late TestGoogleMapsFlutterPlatform platform;
 
   setUp(() {
     // Use a mock platform so we never need to hit the MethodChannel code.
+    platform = TestGoogleMapsFlutterPlatform();
     GoogleMapsFlutterPlatform.instance = platform;
-    resetMockitoState();
-    _setupMock(platform);
   });
 
   testWidgets('_webOnlyMapCreationId increments with each GoogleMap widget', (
@@ -49,18 +49,9 @@ void main() {
     );
 
     // Verify that each one was created with a different _webOnlyMapCreationId.
-    verifyInOrder([
-      platform.buildView(
-        argThat(containsPair('_webOnlyMapCreationId', 0)),
-        any,
-        any,
-      ),
-      platform.buildView(
-        argThat(containsPair('_webOnlyMapCreationId', 1)),
-        any,
-        any,
-      ),
-    ]);
+    expect(platform.createdIds.length, 2);
+    expect(platform.createdIds[0], 0);
+    expect(platform.createdIds[1], 1);
   });
 
   testWidgets('Calls platform.dispose when GoogleMap is disposed of', (
@@ -75,47 +66,220 @@ void main() {
     // Now dispose of the map...
     await tester.pumpWidget(Container());
 
-    verify(platform.dispose(mapId: anyNamed('mapId')));
+    expect(platform.disposed, true);
   });
 }
 
-// Some test setup classes below...
+// A dummy implementation of the platform interface for tests.
+class TestGoogleMapsFlutterPlatform extends GoogleMapsFlutterPlatform {
+  TestGoogleMapsFlutterPlatform();
 
-class _MockStream<T> extends Mock implements Stream<T> {}
+  // The IDs passed to each call to buildView, in call order.
+  List<int> createdIds = <int>[];
 
-typedef _CreationCallback = void Function(int);
+  // Whether `dispose` has been called.
+  bool disposed = false;
 
-// Installs test mocks on the platform
-void _setupMock(MockGoogleMapsFlutterPlatform platform) {
-  // Used to create the view of the map...
-  when(platform.buildView(any, any, any)).thenAnswer((realInvocation) {
-    // Call the onPlatformViewCreated callback so the controller gets created.
-    _CreationCallback onPlatformViewCreatedCb =
-        realInvocation.positionalArguments[2];
-    onPlatformViewCreatedCb.call(0);
+  // Stream controller to inject events for testing.
+  final StreamController<MapEvent> mapEventStreamController =
+      StreamController<MapEvent>.broadcast();
+
+  @override
+  Future<void> init(int mapId) async {}
+
+  @override
+  Future<void> updateMapOptions(
+    Map<String, dynamic> optionsUpdate, {
+    required int mapId,
+  }) async {}
+
+  @override
+  Future<void> updateMarkers(
+    MarkerUpdates markerUpdates, {
+    required int mapId,
+  }) async {}
+
+  @override
+  Future<void> updatePolygons(
+    PolygonUpdates polygonUpdates, {
+    required int mapId,
+  }) async {}
+
+  @override
+  Future<void> updatePolylines(
+    PolylineUpdates polylineUpdates, {
+    required int mapId,
+  }) async {}
+
+  @override
+  Future<void> updateCircles(
+    CircleUpdates circleUpdates, {
+    required int mapId,
+  }) async {}
+
+  @override
+  Future<void> updateTileOverlays({
+    required Set<TileOverlay> newTileOverlays,
+    required int mapId,
+  }) async {}
+
+  @override
+  Future<void> clearTileCache(
+    TileOverlayId tileOverlayId, {
+    required int mapId,
+  }) async {}
+
+  @override
+  Future<void> animateCamera(
+    CameraUpdate cameraUpdate, {
+    required int mapId,
+  }) async {}
+
+  @override
+  Future<void> moveCamera(
+    CameraUpdate cameraUpdate, {
+    required int mapId,
+  }) async {}
+
+  @override
+  Future<void> setMapStyle(
+    String? mapStyle, {
+    required int mapId,
+  }) async {}
+
+  @override
+  Future<LatLngBounds> getVisibleRegion({
+    required int mapId,
+  }) async {
+    return LatLngBounds(southwest: LatLng(0, 0), northeast: LatLng(0, 0));
+  }
+
+  @override
+  Future<ScreenCoordinate> getScreenCoordinate(
+    LatLng latLng, {
+    required int mapId,
+  }) async {
+    return ScreenCoordinate(x: 0, y: 0);
+  }
+
+  @override
+  Future<LatLng> getLatLng(
+    ScreenCoordinate screenCoordinate, {
+    required int mapId,
+  }) async {
+    return LatLng(0, 0);
+  }
+
+  @override
+  Future<void> showMarkerInfoWindow(
+    MarkerId markerId, {
+    required int mapId,
+  }) async {}
+
+  @override
+  Future<void> hideMarkerInfoWindow(
+    MarkerId markerId, {
+    required int mapId,
+  }) async {}
+
+  @override
+  Future<bool> isMarkerInfoWindowShown(
+    MarkerId markerId, {
+    required int mapId,
+  }) async {
+    return false;
+  }
+
+  @override
+  Future<double> getZoomLevel({
+    required int mapId,
+  }) async {
+    return 0.0;
+  }
+
+  @override
+  Future<Uint8List?> takeSnapshot({
+    required int mapId,
+  }) async {
+    return null;
+  }
+
+  @override
+  Stream<CameraMoveStartedEvent> onCameraMoveStarted({required int mapId}) {
+    return mapEventStreamController.stream.whereType<CameraMoveStartedEvent>();
+  }
+
+  @override
+  Stream<CameraMoveEvent> onCameraMove({required int mapId}) {
+    return mapEventStreamController.stream.whereType<CameraMoveEvent>();
+  }
+
+  @override
+  Stream<CameraIdleEvent> onCameraIdle({required int mapId}) {
+    return mapEventStreamController.stream.whereType<CameraIdleEvent>();
+  }
+
+  @override
+  Stream<MarkerTapEvent> onMarkerTap({required int mapId}) {
+    return mapEventStreamController.stream.whereType<MarkerTapEvent>();
+  }
+
+  @override
+  Stream<InfoWindowTapEvent> onInfoWindowTap({required int mapId}) {
+    return mapEventStreamController.stream.whereType<InfoWindowTapEvent>();
+  }
+
+  @override
+  Stream<MarkerDragEndEvent> onMarkerDragEnd({required int mapId}) {
+    return mapEventStreamController.stream.whereType<MarkerDragEndEvent>();
+  }
+
+  @override
+  Stream<PolylineTapEvent> onPolylineTap({required int mapId}) {
+    return mapEventStreamController.stream.whereType<PolylineTapEvent>();
+  }
+
+  @override
+  Stream<PolygonTapEvent> onPolygonTap({required int mapId}) {
+    return mapEventStreamController.stream.whereType<PolygonTapEvent>();
+  }
+
+  @override
+  Stream<CircleTapEvent> onCircleTap({required int mapId}) {
+    return mapEventStreamController.stream.whereType<CircleTapEvent>();
+  }
+
+  @override
+  Stream<MapTapEvent> onTap({required int mapId}) {
+    return mapEventStreamController.stream.whereType<MapTapEvent>();
+  }
+
+  @override
+  Stream<MapLongPressEvent> onLongPress({required int mapId}) {
+    return mapEventStreamController.stream.whereType<MapLongPressEvent>();
+  }
+
+  @override
+  void dispose({required int mapId}) {
+    disposed = true;
+  }
+
+  @override
+  Widget buildView(
+    int creationId,
+    PlatformViewCreatedCallback onPlatformViewCreated, {
+    required CameraPosition initialCameraPosition,
+    Set<Marker> markers = const <Marker>{},
+    Set<Polygon> polygons = const <Polygon>{},
+    Set<Polyline> polylines = const <Polyline>{},
+    Set<Circle> circles = const <Circle>{},
+    Set<TileOverlay> tileOverlays = const <TileOverlay>{},
+    Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers =
+        const <Factory<OneSequenceGestureRecognizer>>{},
+    Map<String, dynamic> mapOptions = const <String, dynamic>{},
+  }) {
+    onPlatformViewCreated(0);
+    createdIds.add(creationId);
     return Container();
-  });
-  // Used to create the Controller
-  when(platform.onCameraIdle(mapId: anyNamed('mapId')))
-      .thenAnswer((_) => _MockStream<CameraIdleEvent>());
-  when(platform.onCameraMove(mapId: anyNamed('mapId')))
-      .thenAnswer((_) => _MockStream<CameraMoveEvent>());
-  when(platform.onCameraMoveStarted(mapId: anyNamed('mapId')))
-      .thenAnswer((_) => _MockStream<CameraMoveStartedEvent>());
-  when(platform.onCircleTap(mapId: anyNamed('mapId')))
-      .thenAnswer((_) => _MockStream<CircleTapEvent>());
-  when(platform.onInfoWindowTap(mapId: anyNamed('mapId')))
-      .thenAnswer((_) => _MockStream<InfoWindowTapEvent>());
-  when(platform.onLongPress(mapId: anyNamed('mapId')))
-      .thenAnswer((_) => _MockStream<MapLongPressEvent>());
-  when(platform.onMarkerDragEnd(mapId: anyNamed('mapId')))
-      .thenAnswer((_) => _MockStream<MarkerDragEndEvent>());
-  when(platform.onMarkerTap(mapId: anyNamed('mapId')))
-      .thenAnswer((_) => _MockStream<MarkerTapEvent>());
-  when(platform.onPolygonTap(mapId: anyNamed('mapId')))
-      .thenAnswer((_) => _MockStream<PolygonTapEvent>());
-  when(platform.onPolylineTap(mapId: anyNamed('mapId')))
-      .thenAnswer((_) => _MockStream<PolylineTapEvent>());
-  when(platform.onTap(mapId: anyNamed('mapId')))
-      .thenAnswer((_) => _MockStream<MapTapEvent>());
+  }
 }
