@@ -9,20 +9,6 @@ import 'package:google_maps_flutter/google_maps_flutter.dart';
 
 import 'fake_maps_controllers.dart';
 
-Set<Marker> _toSet({Marker m1, Marker m2, Marker m3}) {
-  final Set<Marker> res = Set<Marker>.identity();
-  if (m1 != null) {
-    res.add(m1);
-  }
-  if (m2 != null) {
-    res.add(m2);
-  }
-  if (m3 != null) {
-    res.add(m3);
-  }
-  return res;
-}
-
 Widget _mapWithMarkers(Set<Marker> markers) {
   return Directionality(
     textDirection: TextDirection.ltr,
@@ -50,10 +36,10 @@ void main() {
 
   testWidgets('Initializing a marker', (WidgetTester tester) async {
     final Marker m1 = Marker(markerId: MarkerId("marker_1"));
-    await tester.pumpWidget(_mapWithMarkers(_toSet(m1: m1)));
+    await tester.pumpWidget(_mapWithMarkers(<Marker>{m1}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.markersToAdd.length, 1);
 
     final Marker initializedMarker = platformGoogleMap.markersToAdd.first;
@@ -66,11 +52,11 @@ void main() {
     final Marker m1 = Marker(markerId: MarkerId("marker_1"));
     final Marker m2 = Marker(markerId: MarkerId("marker_2"));
 
-    await tester.pumpWidget(_mapWithMarkers(_toSet(m1: m1)));
-    await tester.pumpWidget(_mapWithMarkers(_toSet(m1: m1, m2: m2)));
+    await tester.pumpWidget(_mapWithMarkers(<Marker>{m1}));
+    await tester.pumpWidget(_mapWithMarkers(<Marker>{m1, m2}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.markersToAdd.length, 1);
 
     final Marker addedMarker = platformGoogleMap.markersToAdd.first;
@@ -84,11 +70,11 @@ void main() {
   testWidgets("Removing a marker", (WidgetTester tester) async {
     final Marker m1 = Marker(markerId: MarkerId("marker_1"));
 
-    await tester.pumpWidget(_mapWithMarkers(_toSet(m1: m1)));
-    await tester.pumpWidget(_mapWithMarkers(null));
+    await tester.pumpWidget(_mapWithMarkers(<Marker>{m1}));
+    await tester.pumpWidget(_mapWithMarkers(<Marker>{}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.markerIdsToRemove.length, 1);
     expect(platformGoogleMap.markerIdsToRemove.first, equals(m1.markerId));
 
@@ -100,11 +86,11 @@ void main() {
     final Marker m1 = Marker(markerId: MarkerId("marker_1"));
     final Marker m2 = Marker(markerId: MarkerId("marker_1"), alpha: 0.5);
 
-    await tester.pumpWidget(_mapWithMarkers(_toSet(m1: m1)));
-    await tester.pumpWidget(_mapWithMarkers(_toSet(m1: m2)));
+    await tester.pumpWidget(_mapWithMarkers(<Marker>{m1}));
+    await tester.pumpWidget(_mapWithMarkers(<Marker>{m2}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.markersToChange.length, 1);
     expect(platformGoogleMap.markersToChange.first, equals(m2));
 
@@ -119,11 +105,11 @@ void main() {
       infoWindow: const InfoWindow(snippet: 'changed'),
     );
 
-    await tester.pumpWidget(_mapWithMarkers(_toSet(m1: m1)));
-    await tester.pumpWidget(_mapWithMarkers(_toSet(m1: m2)));
+    await tester.pumpWidget(_mapWithMarkers(<Marker>{m1}));
+    await tester.pumpWidget(_mapWithMarkers(<Marker>{m2}));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
     expect(platformGoogleMap.markersToChange.length, 1);
 
     final Marker update = platformGoogleMap.markersToChange.first;
@@ -134,16 +120,16 @@ void main() {
   testWidgets("Multi Update", (WidgetTester tester) async {
     Marker m1 = Marker(markerId: MarkerId("marker_1"));
     Marker m2 = Marker(markerId: MarkerId("marker_2"));
-    final Set<Marker> prev = _toSet(m1: m1, m2: m2);
+    final Set<Marker> prev = <Marker>{m1, m2};
     m1 = Marker(markerId: MarkerId("marker_1"), visible: false);
     m2 = Marker(markerId: MarkerId("marker_2"), draggable: true);
-    final Set<Marker> cur = _toSet(m1: m1, m2: m2);
+    final Set<Marker> cur = <Marker>{m1, m2};
 
     await tester.pumpWidget(_mapWithMarkers(prev));
     await tester.pumpWidget(_mapWithMarkers(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformGoogleMap.markersToChange, cur);
     expect(platformGoogleMap.markerIdsToRemove.isEmpty, true);
@@ -153,18 +139,18 @@ void main() {
   testWidgets("Multi Update", (WidgetTester tester) async {
     Marker m2 = Marker(markerId: MarkerId("marker_2"));
     final Marker m3 = Marker(markerId: MarkerId("marker_3"));
-    final Set<Marker> prev = _toSet(m2: m2, m3: m3);
+    final Set<Marker> prev = <Marker>{m2, m3};
 
     // m1 is added, m2 is updated, m3 is removed.
     final Marker m1 = Marker(markerId: MarkerId("marker_1"));
     m2 = Marker(markerId: MarkerId("marker_2"), draggable: true);
-    final Set<Marker> cur = _toSet(m1: m1, m2: m2);
+    final Set<Marker> cur = <Marker>{m1, m2};
 
     await tester.pumpWidget(_mapWithMarkers(prev));
     await tester.pumpWidget(_mapWithMarkers(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformGoogleMap.markersToChange.length, 1);
     expect(platformGoogleMap.markersToAdd.length, 1);
@@ -179,35 +165,35 @@ void main() {
     final Marker m1 = Marker(markerId: MarkerId("marker_1"));
     final Marker m2 = Marker(markerId: MarkerId("marker_2"));
     Marker m3 = Marker(markerId: MarkerId("marker_3"));
-    final Set<Marker> prev = _toSet(m1: m1, m2: m2, m3: m3);
+    final Set<Marker> prev = <Marker>{m1, m2, m3};
     m3 = Marker(markerId: MarkerId("marker_3"), draggable: true);
-    final Set<Marker> cur = _toSet(m1: m1, m2: m2, m3: m3);
+    final Set<Marker> cur = <Marker>{m1, m2, m3};
 
     await tester.pumpWidget(_mapWithMarkers(prev));
     await tester.pumpWidget(_mapWithMarkers(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
-    expect(platformGoogleMap.markersToChange, _toSet(m3: m3));
+    expect(platformGoogleMap.markersToChange, <Marker>{m3});
     expect(platformGoogleMap.markerIdsToRemove.isEmpty, true);
     expect(platformGoogleMap.markersToAdd.isEmpty, true);
   });
 
   testWidgets("Update non platform related attr", (WidgetTester tester) async {
     Marker m1 = Marker(markerId: MarkerId("marker_1"));
-    final Set<Marker> prev = _toSet(m1: m1);
+    final Set<Marker> prev = <Marker>{m1};
     m1 = Marker(
         markerId: MarkerId("marker_1"),
         onTap: () => print("hello"),
         onDragEnd: (LatLng latLng) => print(latLng));
-    final Set<Marker> cur = _toSet(m1: m1);
+    final Set<Marker> cur = <Marker>{m1};
 
     await tester.pumpWidget(_mapWithMarkers(prev));
     await tester.pumpWidget(_mapWithMarkers(cur));
 
     final FakePlatformGoogleMap platformGoogleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformGoogleMap.markersToChange.isEmpty, true);
     expect(platformGoogleMap.markerIdsToRemove.isEmpty, true);
